# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School MIS API",
        "description": "High school management information system API built with Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Primary API server"
        }
    ],
    "paths": {
...
        "/api/v1/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "List grade entries with pagination and optional included relationships",
                "description": "Retrieve a paginated list of grade entries. Use 'includes' parameter to eager load related resources (assessment, student).",
                "operationId": "8474a678a878d31a5f633c9448146db0",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of records per page (default 10, max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "includes",
                        "in": "query",
                        "description": "Comma-separated list of relationships to include. Available: assessment, student",
                        "schema": {
                            "type": "string"
                        },
                        "example": "assessment,student"
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Filter by student ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "assessment_id",
                        "in": "query",
                        "description": "Filter by assessment ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grade collection retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedGradeResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Insufficient permissions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Record or update a student grade",
                "operationId": "64ed245e029c1f08d562c87f01f3a5b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GradeStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Grade recorded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GradeResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "schema": "ValidationErrorResponseSchema",
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "UnauthorizedResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "GradeStoreRequest": {
                "required": [
                    "student_id",
                    "assessment_id",
                    "score"
                ],
                "properties": {
                    "student_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "assessment_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "score": {
                        "type": "number",
                        "format": "float"
                    },
                    "remarks": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "GradeResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "student_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "assessment_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "score": {
                        "type": "number",
                        "format": "float"
                    },
                    "grade_letter": {
                        "type": "string"
                    },
                    "remarks": {
                        "type": "string",
                        "nullable": true
                    },
                    "recorded_by": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "recorded_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "student": {
                        "type": "object"
                    },
                    "assessment": {
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "format": "uri",
                        "example": "/api/v1/students"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "format": "uri",
                        "example": "/api/v1/students?page=1"
                    },
                    "last": {
                        "type": "string",
                        "format": "uri",
                        "example": "/api/v1/students?page=10"
                    },
                    "prev": {
                        "type": "string",
                        "format": "uri",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "format": "uri",
                        "example": "/api/v1/students?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginatedGradeResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GradeResource"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}